+++ geo/CHANGES.md
@@ -19,6 +19,8 @@
 - `Intersections` no longer `panic`'s when given pathological input (and it's typically much faster!)
   - <https://github.com/georust/geo/pull/1387>
   - <https://github.com/georust/geo/pull/1358>
+- BREAKING: The `Crosses` trait used by `Intersections` now returns a `Line`, not a `LineOrPoint`.
+  - <https://github.com/georust/geo/pull/1359>
 - Added: Geometry buffering to "grow" or "shrink" a geometry by creating a buffer whose boundary is the specified offset from the input.
   - <https://github.com/georust/geo/pull/1365>
 - BREAKING: `BoolOpsNum` must now implement GeoFloat, not just GeoNum. In practice, this shouldn't break for any concrete types (like f32, f64).